Dark mode style changes

+++ CampKit/Views/EditableItemView.swift
@@ -29,16 +29,18 @@ struct EditableItemView: View {
             // Background for delete action
             HStack {
                 Spacer()
-                Rectangle()
-                    .fill(Color.red)
-                    .frame(width: abs(offset)) // Show background width based on drag distance
-                    .overlay(
-                        Image(systemName: "trash")
-                            .foregroundColor(.white)
-                            .font(.system(size: trashIconSize()))
-                            .padding(.trailing, 16),
-                        alignment: .trailing
-                    )
+                if offset != 0 {
+                    Rectangle()
+                        .fill(Color.red)
+                        .frame(width: abs(offset)) // Show background width based on drag distance
+                        .overlay(
+                            Image(systemName: "trash")
+                                .foregroundColor(.white)
+                                .font(.system(size: trashIconSize()))
+                                .padding(.trailing, 16),
+                            alignment: .trailing
+                        )
+                }
             }
             
             // Foreground for list item
@@ -60,8 +62,7 @@ struct EditableItemView: View {
             .offset(x: offset)
             .padding(.vertical, 5)
             .padding(.horizontal)
-            .background(Color.white) // Prevent bleed-through from red background
-            .cornerRadius(5)
+            .background(Color.colorWhite) // Prevent bleed-through from red background
             .offset(x: offset)
             .gesture(
                 DragGesture()
@@ -79,23 +80,30 @@ struct EditableItemView: View {
                     }
                     .onEnded { _ in
                         if willDelete {
-                            deleteItem()
-                            hapticFeedback.notificationOccurred(.success)
+                            if abs(offset) > deletionThreshold {
+                                withAnimation {
+                                    offset = -UIScreen.main.bounds.width
+                                }
+                                DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
+                                    deleteItem()
+                                    hapticFeedback.notificationOccurred(.success)
+                                }
+                            }
                         } else {
-                            withAnimation {
+                            withAnimation(.spring) {
                                 offset = 0 // Reset position if not deleted
                             }
                         }
                     }
             )
-            .animation(.spring(), value: offset)
+           .animation(.spring(), value: offset)
         }//:ZSTACK
         
     }//:BODY
     
     private func trashIconSize() -> CGFloat {
-        let baseSize: CGFloat = 18
-        let maxSize: CGFloat = 24
+        let baseSize: CGFloat = 14
+        let maxSize: CGFloat = 20
         let scaleFactor = min(abs(offset) / deletionThreshold, 1.0) // Max scale is 1.0
         return baseSize + (maxSize - baseSize) * scaleFactor
     }
